Refactor quantity handling to use decimal values across the application
- Updated quantity fields in models to use float instead of integer for better precision.
- Changed database migration to alter relevant columns from integer to decimal(15, 2).
- Adjusted form inputs and calculations in views to handle decimal values correctly.
- Enhanced validation logic to accommodate decimal quantities.
- Modified display formatting for quantities in various views to show two decimal places.
- Ensured consistent handling of stock quantities in services and controllers.

+++ app/Models/DetailPenjualan.php
@@ -23,7 +23,7 @@ class DetailPenjualan extends Model
     ];
 
     protected $casts = [
-        'qty' => 'integer',
+        'qty' => 'float',
         'harga_awal' => 'decimal:2',
         'harga_jual' => 'decimal:2',
         'subtotal' => 'decimal:2',
